Add files via upload
It is a game of scrabble built in c language  .It takes an input from two players, player 1 and player 2. Both have to type a word. Each letter has a specific number assigned to it. The score of each player is the sum of the numbers assigned to its letters. The one who has the higher score wins. Here I have used the concept of array to make things easier.